+++ meridian/analysis/review/results_test.py
@@ -42,9 +42,11 @@ class ConvergenceCheckResultTest(parameterized.TestCase):
     result = results.ConvergenceCheckResult(
         case=results.ConvergenceCases.CONVERGED,
         config=config,
-        max_rhat=1.0,
+        max_r_hat=1.0,
         max_parameter="mock_var",
     )
+    self.assertEqual(result.max_r_hat, 1.0)
+    self.assertEqual(result.max_rhat, 1.0)
     self.assertEqual(result.case.status, results.Status.PASS)
     self.assertEqual(
         result.recommendation,
@@ -57,7 +59,7 @@ class ConvergenceCheckResultTest(parameterized.TestCase):
     result = results.ConvergenceCheckResult(
         case=results.ConvergenceCases.NOT_FULLY_CONVERGED,
         config=config,
-        max_rhat=3.0,
+        max_r_hat=3.0,
         max_parameter="mock_var",
     )
     self.assertEqual(result.case.status, results.Status.FAIL)
@@ -73,7 +75,7 @@ class ConvergenceCheckResultTest(parameterized.TestCase):
     result = results.ConvergenceCheckResult(
         case=results.ConvergenceCases.NOT_CONVERGED,
         config=config,
-        max_rhat=11.0,
+        max_r_hat=11.0,
         max_parameter="mock_var",
     )
     self.assertEqual(result.case.status, results.Status.FAIL)
@@ -534,7 +536,7 @@ class ReviewSummaryTest(parameterized.TestCase):
     mock_result = results.ConvergenceCheckResult(
         case=results.ConvergenceCases.CONVERGED,
         config=configs.ConvergenceConfig(),
-        max_rhat=1.0,
+        max_r_hat=1.0,
         max_parameter="mock_var",
     )
     summary = results.ReviewSummary(